Move EventLoopClock into EventLoop and rename it to Clock
