--- concurrent/actor/EventLoopClock.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (C) 2020 Grakn Labs
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package grakn.common.concurrent.actor;
-
-import java.util.Random;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.function.Supplier;
-
-public class EventLoopClock {
-    private static Supplier<Long> getTime = () -> java.lang.System.currentTimeMillis();
-    private static Random random = ThreadLocalRandom.current();
-
-    private EventLoopClock() {}
-
-    public static void set(Supplier<Long> getTime, Random random) {
-        EventLoopClock.getTime = getTime;
-        EventLoopClock.random = random;
-    }
-
-    public static long time() {
-        return getTime.get();
-    }
-
-    public static Random random() {
-        return random;
-    }
-}
